requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
Django==5.0.7
django-environ==0.11.2
djangorestframework==3.15.2
drf-yasg==1.21.7
h11==0.14.0
inflection==0.5.1
mysqlclient==2.2.4
packaging==24.1
pytz==2024.1
PyYAML==6.0.1
setuptools==72.1.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
uvicorn==0.30.3
